runner.docker: Pass through all NEXTSTRAIN_* environment variables
This reduces the specific coupling of pathogen builds to the CLI by
instituting a generic convention all builds can use.  It very
intentionally is a solution that will work seamlessly with all build
execution methods (e.g. locally with snakemake or via nextstrain build)
so that our instructions can remain consistent and unified.

One downside is that _how_ environment variables get set in the first
place is still left up to each individual user.  There is not a standard
way, although environment directories remain an option even if they're
not integrated.  We can integrate envdir or something similarly
prescriptive later if we still desire it.

This approach came out of summarizing my thoughts around environment
vars and the CLI and pathogen builds a couple weeks ago in this
document:

   https://github.com/tsibley/blab-standup/blob/master/2018-07-20.md#20-july-2018

A little bit of further discussion, linked to from the top of that
document, was had on Slack in #nextstrain-general.

I ultimately chose the prefix convention as its a simpler step forward
and has the advantage of requiring no custom boilerplate to create an
environment directory in remote execution environments like Travis CI
and AWS Batch.